requirements.txt
# See LICENSE for license information.

GitPython
